Complete 4.21

+++ part4/blog/controllers/blogs.js
@@ -59,6 +59,25 @@ blogRouter.get('/:id', async (request, response, next) => {
 })
 
 blogRouter.delete('/:id', async (request, response, next) => {
+
+  let decodedToken
+
+  try {
+    decodedToken = jwt.verify(request.token, process.env.SECRET)
+  } catch (err) {
+    next(err)
+  }
+
+  if (!request.token || !decodedToken.id) {    
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() !== decodedToken.id) {
+    return response.status(401).json({ error: 'invalid creditentials'})
+  }
+
   try {
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
